wrote code for button bindings and commands (also debugged some things with errors)

+++ src/main/java/frc/robot/subsystems/climb/ClimbConstants.java
@@ -1,9 +1,11 @@
 package frc.robot.subsystems.climb;
 
 public class ClimbConstants {
+  // CHANGE LATER
   public static final double voltage = 1;
   public static final int motorPort = 1;
   public static final int CurrentLimit = 0;
-  public static final double moi = 0;
-  public static final double gearing = 0;
+
+  public static final int moi = 0;
+  public static final int gearing = 0;
 }
